Create assembly for step module
